Docker: Fixes to build script and setup script

+++ scripts/base/build.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 
-git clone https://github.com/alessandrolaferlita/RNAdetector.git
+if ! git clone https://github.com/alessandrolaferlita/RNAdetector.git; then
+    exit 1
+fi
 tar -zcvf repo.tar.gz --owner=0 --group=0 ./RNAdetector
 rm -rf ./RNAdetector
 docker build -t alaimos/ubuntu-private:RNAdetector.v1.0 .
